refactor: 전체 프로젝트 구조 및 불필요한 로직 리팩터링
* 로컬 Elasticsearch, Kibana 7.6.1 환경 셋팅

* 개발환경 local, dev, prod 로 분리

* 기존 DTO -> inner class DTO로 변경

* API별 URI versioning 제거

* 불필요한 상수(final) 제거

* 서비스 레이어 인터페이스 제거, 현재 구조에는 장점이 없다 생각
