+++ build.gradle
@@ -22,13 +22,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'com.fasterxml.jackson.core:jackson-core:2.12.3'
-
-	// https://mvnrepository.com/artifact/org.apache.commons/commons-lang3
-	implementation 'org.apache.commons:commons-lang3:3.12.0'
-
-	// https://mvnrepository.com/artifact/javax.validation/validation-api
-	implementation 'javax.validation:validation-api:2.0.1.Final'
+	implementation 'com.fasterxml.jackson.core:jackson-core'
+	implementation 'org.apache.commons:commons-lang3'
+	implementation 'javax.validation:validation-api'
 
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
