+++ src/main/java/com/elasticsearch/controller/IndexController.java
@@ -2,32 +2,27 @@ package com.elasticsearch.controller;
 
 import com.elasticsearch.service.IndexService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-/**
- * Elasticsearch의 Index를 생성, 수정, 삭제하기위한 컨트롤러입니다.
- *
- * @author ymkim
- * @since 2022.03.05 Sat 17:38
- */
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api")
 @RequiredArgsConstructor
+@Slf4j
 public class IndexController {
-    private final Logger log = LoggerFactory.getLogger(IndexController.class);
 
     private final IndexService indexService;
 
     @GetMapping("/index")
-    public ResponseEntity<String> recreateAllIndices(@RequestBody final Map<String, Object> paramMap) {
+    public ResponseEntity<String> recreateAllIndices(@RequestBody Map<String, Object> paramMap) {
         log.info("paramMap.toString = {}", paramMap.toString());
         Map<String, Object> requestParamMap = new LinkedHashMap<>();
         requestParamMap.put("isRecreateIndexYn", true);
