+++ src/main/java/com/elasticsearch/controller/TestController.java
@@ -1,21 +1,18 @@
 package com.elasticsearch.controller;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.swing.*;
-import java.util.Map;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api")
+@Slf4j
 public class TestController {
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @GetMapping(value = "/test")
     public String intro(@RequestParam String title) {
-        logger.info("title = {}", title);
         String introduction = "Welcome to " + title + " intro.";
         return introduction;
     }
