+++ src/main/java/com/elasticsearch/controller/VehicleController.java
@@ -1,13 +1,12 @@
 package com.elasticsearch.controller;
 
 import com.elasticsearch.document.Vehicle;
-import com.elasticsearch.dto.request.VehicleRequestDto;
+import com.elasticsearch.dto.VehicleRequest;
 import com.elasticsearch.search.SearchRequestDto;
 import com.elasticsearch.service.VehicleService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.time.StopWatch;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,33 +16,22 @@ import java.util.Date;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-/**
- * 차량 정보 관련 검색 컨트롤러 입니다.
- *
- * @author ymkim
- * @since 2022.03.05 Sat 13:15
- */
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api")
 @RequiredArgsConstructor
+@Slf4j
 public class VehicleController {
-    private final Logger log = LoggerFactory.getLogger(VehicleController.class);
 
     private final VehicleService vehicleService;
 
-    /**
-     * Vehicle Document 생성
-     *
-     * @param vehicleRequestDto
-     */
     @PostMapping(value = "/vehicle")
-    public void saveDocumentToIndex(@RequestBody final VehicleRequestDto vehicleRequestDto) {
+    public void saveDocumentToIndex(@RequestBody VehicleRequest.request vehicleRequestDto) {
         log.info("vehicle = {}", vehicleRequestDto.toString());
         vehicleService.saveDocumentToIndex(vehicleRequestDto);
     }
 
     @GetMapping(value = "/vehicle/{id}")
-    public ResponseEntity<Vehicle> findVehicleById(@PathVariable final String id) {
+    public ResponseEntity<Vehicle> findVehicleById(@PathVariable String id) {
         log.info("id = {}", id);
         return new ResponseEntity<>(vehicleService.findVehicleById(id), HttpStatus.OK);
     }
@@ -60,22 +48,21 @@ public class VehicleController {
     }
 
     @PostMapping("/vehicle/search")
-    public List<Vehicle> searchVehiclesByMatchQuery(@RequestBody final SearchRequestDto searchRequestDto) {
+    public List<Vehicle> searchVehiclesByMatchQuery(@RequestBody SearchRequestDto searchRequestDto) {
         log.info("searchRequestDto = {}", searchRequestDto.toString());
         return vehicleService.searchVehiclesByMatchQuery(searchRequestDto);
     }
 
     @GetMapping(value = "/vehicle/search/{date}")
-    public List<Vehicle> searchVehiclesByRangeQuery(
-        @PathVariable @DateTimeFormat(pattern = "yyyy-MM-dd") final Date date) {
+    public List<Vehicle> searchVehiclesByRangeQuery(@PathVariable @DateTimeFormat(pattern = "yyyy-MM-dd") Date date) {
         log.info("date = {}", date.toString());
         return vehicleService.searchVehiclesByRangeQuery(date);
     }
 
     @PostMapping(value = "/vehicle/search/{date}")
     public List<Vehicle> searchVehiclesByBoolQuery(
-        @RequestBody final SearchRequestDto searchRequestDto,
-        @PathVariable @DateTimeFormat(pattern = "yyyy-MM-dd") final Date date) {
+            @RequestBody SearchRequestDto searchRequestDto,
+            @PathVariable @DateTimeFormat(pattern = "yyyy-MM-dd") Date date) {
         log.info("searchRequestDto = {}", searchRequestDto.toString());
         return vehicleService.searchVehiclesByBoolQuery(searchRequestDto, date);
     }
