+++ src/main/java/com/elasticsearch/helper/IndicesHelper.java
@@ -1,11 +1,5 @@
 package com.elasticsearch.helper;
 
-/**
- * 인덱스명을 따로 관리하기 위한 클래스입니다.
- *
- * @author ymkim
- * @since 2022.03.05 Sat 14:23
- */
 public final class IndicesHelper {
 
     public static final String USER_INDEX_NAME = "user";
