+++ src/main/java/com/elasticsearch/search/SearchRequestDto.java
@@ -10,8 +10,10 @@ import java.util.List;
 @ToString
 @NoArgsConstructor
 public class SearchRequestDto extends PagedRequestDTO {
+
     private List<String> fields;    // 검색을 원하는 필드명
     private String searchTerm;      // 검색 단어(Term 단위)
     private String sortBy;          // 정렬 기준 필드, ex) created_at
     private SortOrder order;        // 정렬 기준, ex) DESC, ASC
+
 }
