+++ src/main/resources/application-dev.yml
@@ -1,0 +1,10 @@
+server:
+  port: 8082
+
+spring:
+  config:
+    activate:
+      on-profile: dev
+  jpa:
+    hibernate:
+      ddl-auto: update
