+++ src/main/resources/application-local.yml
@@ -1,0 +1,10 @@
+server:
+  port: 8081
+
+spring:
+  config:
+    activate:
+      on-profile: local
+  jpa:
+    hibernate:
+      ddl-auto: create
