+++ src/main/resources/application-prod.yml
@@ -1,0 +1,10 @@
+server:
+  port: 8083
+
+spring:
+  config:
+    activate:
+      on-profile: prod
+  jpa:
+    hibernate:
+      ddl-auto: none
